requirements.txt
streamlit==1.42.2
streamlit-drawable-canvas==0.9.3
torch==2.6.0
torchvision==0.21.0
pillow==11.1.0
numpy==1.26.4
opencv-python==4.11.0.86
matplotlib==3.10.1